backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2026.1.4",
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "google-generativeai>=0.8.6",
    "pandas>=3.0.0",
    "pymongo>=4.16.0",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
